Collection Completed
1. Collection Loan Details -Completed.
2. collection submit - Completed.
3. Due chart, Penalty chart, fine chart - Completed.
4. Fine add - Completed.
5. Updated database added.

+++ api/bank_creation/delete_bank_creation.php
@@ -7,7 +7,9 @@ $qry = $pdo->prepare("SELECT qr_code FROM bank_creation WHERE id = $id ");
 if ($qry->execute()) {
     $row = $qry->fetch(PDO::FETCH_ASSOC);
 
-    unlink("../../uploads/bank_creation/qr_code/" . $row['qr_code']);
+    if($row['qr_code'] !=''){
+        unlink("../../uploads/bank_creation/qr_code/" . $row['qr_code']);
+    }
 
     $qry = $pdo->query("DELETE FROM `bank_creation` WHERE `id` = '$id'");
 }
